Added working version of Mad Science based on current MineAPI #53. This build is for developers only do not use for playing!

+++ src/main/java/universalelectricity/api/item/ItemElectric.java
@@ -138,7 +138,7 @@ public abstract class ItemElectric extends Item implements IEnergyItem, IVoltage
             }
             else if (obj instanceof NBTTagLong)
             {
-                energyStored = (long) ((NBTTagLong) obj).data;
+                energyStored = ((NBTTagLong) obj).data;
             }
         }
